requirements.txt
networkx==3.0
pyemd==1.0.0
geomloss==0.2.6
pandas==2.0.2
tensorboard
numpy
scipy
tqdm
ipython
matplotlib
pyyaml
torchtyping
pot
pytest
torchdyn
lightning
comet_ml
vector
uproot
h5py
fastjet
seaborn
rich
